requirements.txt
alembic==1.12.1
asgiref==3.7.2
asttokens==2.3.0
autopep8==2.0.4
backcall==0.2.0
bcrypt==4.0.1
blinker==1.6.2
cachelib==0.10.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
colorgram.py==1.2.0
comm==0.1.4
contourpy==1.2.1
cycler==0.12.1
debugpy==1.6.7.post1
decorator==5.1.1
Django==4.2.6
dnspython==2.4.2
docopt==0.6.2
email-validator==2.1.0.post1
executing==1.2.0
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-Migrate==4.0.5
Flask-OAuthlib==0.9.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.1.1
fonttools==4.51.0
google-api-core==2.12.0
google-api-python-client==2.105.0
google-auth==2.23.1
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.61.0
greenlet==3.0.1
gspread==5.11.2
httplib2==0.22.0
idna==3.4
ipykernel==6.25.1
ipython==8.15.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
jupyter_client==8.3.1
jupyter_core==5.3.1
kiwisolver==1.4.5
Mako==1.3.0
MarkupSafe==2.1.3
matplotlib==3.9.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.7
numpy==1.26.0
oauthlib==2.1.0
packaging==23.1
pandas==2.1.1
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.1
pipreqs==0.4.13
platformdirs==3.10.0
plotly==5.22.0
prettytable==3.9.0
prompt-toolkit==3.0.39
protobuf==4.24.4
psutil==5.9.5
psycopg2-binary==2.9.9
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.11.0
Pygments==2.16.1
pyparsing==3.1.1
pyperclip==1.8.2
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.1.0
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.22
sqlparse==0.4.4
stack-data==0.6.2
tenacity==8.3.0
tornado==6.3.3
traitlets==5.9.0
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.5
wcwidth==0.2.6
Werkzeug==2.3.7
WTForms==3.0.1
yarg==0.1.9
pytest
